Added the code to do graph colouring and graph partitioning
Added the code with the time metrics to check how the distance metrics lead to different graph colours and their communities. Also added time metrics to check the bottleneck in the plotting of graph colouring and graph partitioning of these different metrics for different resolutions ranging from 10x10 to 50x50 and plotting the results for the same.